Allow shrinking `Window` size with `fixed_size()`
Without this, if you've increased a Window size by setting
`fixed_size` and then tried to go back to a smaller value, it would
not work, because the previous size would be loaded from memory.

I think there's probably a better way to do this (we need to
understand better why it's loading the value from memory instead of
just using the new size), but this works for now.